clang-tidy: integrate

+++ include/lart/taint.hpp
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <cstdint>
 #include <sanitizer/dfsan_interface.h>
 
 namespace __lart::runtime
@@ -18,7 +19,7 @@ namespace __lart::runtime
 
     template< typename integral > bool is_tainted( integral value )
     {
-        auto label = dfsan_get_label( value );
+        auto label = dfsan_get_label( static_cast< std::int64_t >( value ) );
         return dfsan_has_label( label, taint );
     }
